+++ changelog.md
@@ -1,4 +1,9 @@
 # Changelog
+## adminer-dba-tool 0.1.1 (06.06.2022)
+- Update new_files/admin/adminer/adminer.php von v4.7.9 auf v4.8.1
+- Update docs/install.md Text in Deinstallation ergänzt.
+- Update README.md Text ergänzt.
+- Update mehrere Dateien im Ordner new_files zur Kompatibilität mit PHP 8 angepasst. 
 ## adminer-dba-tool 0.1.1 (08.02.2021)
 - Update new_files/admin/adminer/adminer.php von v4.7.8 auf v4.7.9
 - Update docs/install.md Text in Deinstallation ergänzt.